@@ -1,5 +1,5 @@
-// flow-typed signature: 2be8d574e9f7030854fc60dac99e21af
-// flow-typed version: <<STUB>>/babel-cli_v^6.24.1/flow_v0.48.0
+// flow-typed signature: 8b8a2c739d679c366c7882ad5c42fcad
+// flow-typed version: <<STUB>>/babel-cli_v^6.24.1/flow_v0.53.1
 
 /**
  * This is an autogenerated libdef stub for:
