@@ -1,5 +1,5 @@
-// flow-typed signature: fb9050b753124adf69b920326736cddb
-// flow-typed version: <<STUB>>/babel-eslint_v^7.2.3/flow_v0.48.0
+// flow-typed signature: 897d7e27ab5c9ea597e798bc63bc3129
+// flow-typed version: <<STUB>>/babel-eslint_v^7.2.3/flow_v0.53.1
 
 /**
  * This is an autogenerated libdef stub for:
