@@ -1,5 +1,5 @@
-// flow-typed signature: 66c62e39652763a02af7066f6d54a45e
-// flow-typed version: <<STUB>>/babel-generator_v^6.25.0/flow_v0.48.0
+// flow-typed signature: e00d174c2224b5baa46ea200a7a69e2e
+// flow-typed version: <<STUB>>/babel-generator_v^6.25.0/flow_v0.53.1
 
 /**
  * This is an autogenerated libdef stub for:
