@@ -1,5 +1,5 @@
-// flow-typed signature: 9513ece6a38cfbf075893a25afcdcf62
-// flow-typed version: <<STUB>>/babel-jest_v^20.0.3/flow_v0.48.0
+// flow-typed signature: 7fbd18e47e19516448cc5cb80acdd94f
+// flow-typed version: <<STUB>>/babel-jest_v^20.0.3/flow_v0.53.1
 
 /**
  * This is an autogenerated libdef stub for:
