@@ -1,5 +1,5 @@
-// flow-typed signature: 496797fba4c60e6d4a9608eb2cce32bd
-// flow-typed version: <<STUB>>/babel-plugin-external-helpers_v^6.22.0/flow_v0.48.0
+// flow-typed signature: db5c2f50905a0d07524797ba00d2e27e
+// flow-typed version: <<STUB>>/babel-plugin-external-helpers_v^6.22.0/flow_v0.53.1
 
 /**
  * This is an autogenerated libdef stub for:
