@@ -1,5 +1,5 @@
-// flow-typed signature: 424ac8a5e668e71c8f5c132d85caa2d9
-// flow-typed version: <<STUB>>/babel-plugin-syntax-jsx_v^6.18.0/flow_v0.48.0
+// flow-typed signature: 82413ce8caadf0562e58a7e4c6fecd50
+// flow-typed version: <<STUB>>/babel-plugin-syntax-jsx_v^6.18.0/flow_v0.53.1
 
 /**
  * This is an autogenerated libdef stub for:
