@@ -1,5 +1,5 @@
-// flow-typed signature: 289f574cd48e3bdda0d110bb28cddce5
-// flow-typed version: <<STUB>>/babel-plugin-transform-class-properties_v^6.24.1/flow_v0.48.0
+// flow-typed signature: e63aad8e22191503bc4f41246e2669e7
+// flow-typed version: <<STUB>>/babel-plugin-transform-class-properties_v^6.24.1/flow_v0.53.1
 
 /**
  * This is an autogenerated libdef stub for:
