@@ -1,5 +1,5 @@
-// flow-typed signature: df61fdfe2a941bb0a1a58469cfe8219c
-// flow-typed version: <<STUB>>/babel-plugin-transform-es2015-modules-commonjs_v^6.24.1/flow_v0.48.0
+// flow-typed signature: af46f76461b4337ce21a893449bc992a
+// flow-typed version: <<STUB>>/babel-plugin-transform-es2015-modules-commonjs_v^6.24.1/flow_v0.53.1
 
 /**
  * This is an autogenerated libdef stub for:
