@@ -1,5 +1,5 @@
-// flow-typed signature: 68032a0f1edfd9671ff589311bf27cda
-// flow-typed version: <<STUB>>/babel-plugin-transform-object-rest-spread_v^6.23.0/flow_v0.48.0
+// flow-typed signature: cf492b43d0f4b5b900b619fe25aaa2fa
+// flow-typed version: <<STUB>>/babel-plugin-transform-object-rest-spread_v^6.23.0/flow_v0.53.1
 
 /**
  * This is an autogenerated libdef stub for:
