@@ -1,5 +1,5 @@
-// flow-typed signature: 26d90c22d51781407eb7c1802f2b9659
-// flow-typed version: <<STUB>>/babel-plugin-transform-remove-strict-mode_v^0.0.2/flow_v0.48.0
+// flow-typed signature: f0f52ddd8dabbebb7a7e05a3430df55e
+// flow-typed version: <<STUB>>/babel-plugin-transform-remove-strict-mode_v^0.0.2/flow_v0.53.1
 
 /**
  * This is an autogenerated libdef stub for:
