@@ -1,5 +1,5 @@
-// flow-typed signature: 2004dc8aa54efd841cbeb64749ac7fc8
-// flow-typed version: <<STUB>>/babel-preset-env_v^1.5.2/flow_v0.48.0
+// flow-typed signature: 6218952fe9282dfd826a6ae756cdbc19
+// flow-typed version: <<STUB>>/babel-preset-env_v^1.6.0/flow_v0.53.1
 
 /**
  * This is an autogenerated libdef stub for:
