@@ -1,5 +1,5 @@
-// flow-typed signature: ebb227eb2063cd511833338670db0d5e
-// flow-typed version: <<STUB>>/babel-preset-react_v^6.24.1/flow_v0.48.0
+// flow-typed signature: 2bef25f2440aeb196f7cb9094e88a27f
+// flow-typed version: <<STUB>>/babel-preset-react_v^6.24.1/flow_v0.53.1
 
 /**
  * This is an autogenerated libdef stub for:
