@@ -1,5 +1,5 @@
-// flow-typed signature: 52065904af7a0c1a16f0e95737f35684
-// flow-typed version: <<STUB>>/enzyme-to-json_v^1.5.1/flow_v0.48.0
+// flow-typed signature: 723960cb754223759bd8f18bb80a9b05
+// flow-typed version: <<STUB>>/enzyme-to-json_v^1.5.1/flow_v0.53.1
 
 /**
  * This is an autogenerated libdef stub for:
