@@ -1,5 +1,5 @@
-// flow-typed signature: 8b20232cf5b0dff77e68cd097b2fc2b5
-// flow-typed version: <<STUB>>/eslint-config-standard_v^10.2.1/flow_v0.48.0
+// flow-typed signature: babe78034db8a3de1c77c74123935401
+// flow-typed version: <<STUB>>/eslint-config-standard_v^10.2.1/flow_v0.53.1
 
 /**
  * This is an autogenerated libdef stub for:
