@@ -1,5 +1,5 @@
-// flow-typed signature: 7aeb375c63190086c3e5d9bf041a0ad2
-// flow-typed version: <<STUB>>/eslint-plugin-flowtype_v^2.34.0/flow_v0.48.0
+// flow-typed signature: a61801318a6316556efaf77b3c17a45a
+// flow-typed version: <<STUB>>/eslint-plugin-flowtype_v^2.34.0/flow_v0.53.1
 
 /**
  * This is an autogenerated libdef stub for:
