@@ -1,5 +1,5 @@
-// flow-typed signature: b6631939a281b271c6955707fc3d10b1
-// flow-typed version: <<STUB>>/eslint-plugin-import_v^2.3.0/flow_v0.48.0
+// flow-typed signature: d0885abb34f3e3b438871b111ff2d5e1
+// flow-typed version: <<STUB>>/eslint-plugin-import_v^2.7.0/flow_v0.53.1
 
 /**
  * This is an autogenerated libdef stub for:
