@@ -1,5 +1,5 @@
-// flow-typed signature: 4ef041934d82835ed7a470d29e328704
-// flow-typed version: <<STUB>>/eslint-plugin-node_v^5.0.0/flow_v0.48.0
+// flow-typed signature: 06129166ec619011a95330e384d7274f
+// flow-typed version: <<STUB>>/eslint-plugin-node_v^5.0.0/flow_v0.53.1
 
 /**
  * This is an autogenerated libdef stub for:
@@ -134,6 +134,10 @@ declare module 'eslint-plugin-node/lib/util/get-require-targets' {
   declare module.exports: any;
 }
 
+declare module 'eslint-plugin-node/lib/util/get-resolve-paths' {
+  declare module.exports: any;
+}
+
 declare module 'eslint-plugin-node/lib/util/get-try-extensions' {
   declare module.exports: any;
 }
@@ -234,6 +238,9 @@ declare module 'eslint-plugin-node/lib/util/get-package-json.js' {
 }
 declare module 'eslint-plugin-node/lib/util/get-require-targets.js' {
   declare module.exports: $Exports<'eslint-plugin-node/lib/util/get-require-targets'>;
+}
+declare module 'eslint-plugin-node/lib/util/get-resolve-paths.js' {
+  declare module.exports: $Exports<'eslint-plugin-node/lib/util/get-resolve-paths'>;
 }
 declare module 'eslint-plugin-node/lib/util/get-try-extensions.js' {
   declare module.exports: $Exports<'eslint-plugin-node/lib/util/get-try-extensions'>;
