@@ -1,5 +1,5 @@
-// flow-typed signature: b6cdee81f2887f8c013f91d182f2b249
-// flow-typed version: <<STUB>>/eslint-plugin-promise_v^3.5.0/flow_v0.48.0
+// flow-typed signature: 3d0aaab25f54d91d51e12553286b5694
+// flow-typed version: <<STUB>>/eslint-plugin-promise_v^3.5.0/flow_v0.53.1
 
 /**
  * This is an autogenerated libdef stub for:
