@@ -1,5 +1,5 @@
-// flow-typed signature: f24d4916476210f104f3afc04d7aa43a
-// flow-typed version: <<STUB>>/eslint-plugin-standard_v^3.0.1/flow_v0.48.0
+// flow-typed signature: d097f15532ab8c76c8fa6a3ac68c9c27
+// flow-typed version: <<STUB>>/eslint-plugin-standard_v^3.0.1/flow_v0.53.1
 
 /**
  * This is an autogenerated libdef stub for:
