@@ -1,5 +1,5 @@
-// flow-typed signature: 6256d34ba4171e845c606f6d4251b90d
-// flow-typed version: <<STUB>>/eslint_v^4.0.0/flow_v0.48.0
+// flow-typed signature: f2dd4f6b05d0425382f095a25f0205b2
+// flow-typed version: <<STUB>>/eslint_v^4.0.0/flow_v0.53.1
 
 /**
  * This is an autogenerated libdef stub for:
@@ -23,6 +23,10 @@ declare module 'eslint' {
  * needed.
  */
 declare module 'eslint/bin/eslint' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/conf/config-schema' {
   declare module.exports: any;
 }
 
@@ -95,6 +99,10 @@ declare module 'eslint/lib/config' {
 }
 
 declare module 'eslint/lib/config/autoconfig' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/config/config-cache' {
   declare module.exports: any;
 }
 
@@ -351,6 +359,10 @@ declare module 'eslint/lib/rules/func-style' {
 }
 
 declare module 'eslint/lib/rules/generator-star-spacing' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/rules/getter-return' {
   declare module.exports: any;
 }
 
@@ -1230,6 +1242,10 @@ declare module 'eslint/lib/testers/rule-tester' {
   declare module.exports: any;
 }
 
+declare module 'eslint/lib/testers/test-parser' {
+  declare module.exports: any;
+}
+
 declare module 'eslint/lib/timing' {
   declare module.exports: any;
 }
@@ -1283,6 +1299,10 @@ declare module 'eslint/lib/token-store/skip-cursor' {
 }
 
 declare module 'eslint/lib/token-store/utils' {
+  declare module.exports: any;
+}
+
+declare module 'eslint/lib/util/ajv' {
   declare module.exports: any;
 }
 
@@ -1354,6 +1374,9 @@ declare module 'eslint/lib/util/xml-escape' {
 declare module 'eslint/bin/eslint.js' {
   declare module.exports: $Exports<'eslint/bin/eslint'>;
 }
+declare module 'eslint/conf/config-schema.js' {
+  declare module.exports: $Exports<'eslint/conf/config-schema'>;
+}
 declare module 'eslint/conf/default-cli-options.js' {
   declare module.exports: $Exports<'eslint/conf/default-cli-options'>;
 }
@@ -1407,6 +1430,9 @@ declare module 'eslint/lib/config.js' {
 }
 declare module 'eslint/lib/config/autoconfig.js' {
   declare module.exports: $Exports<'eslint/lib/config/autoconfig'>;
+}
+declare module 'eslint/lib/config/config-cache.js' {
+  declare module.exports: $Exports<'eslint/lib/config/config-cache'>;
 }
 declare module 'eslint/lib/config/config-file.js' {
   declare module.exports: $Exports<'eslint/lib/config/config-file'>;
@@ -1599,6 +1625,9 @@ declare module 'eslint/lib/rules/func-style.js' {
 }
 declare module 'eslint/lib/rules/generator-star-spacing.js' {
   declare module.exports: $Exports<'eslint/lib/rules/generator-star-spacing'>;
+}
+declare module 'eslint/lib/rules/getter-return.js' {
+  declare module.exports: $Exports<'eslint/lib/rules/getter-return'>;
 }
 declare module 'eslint/lib/rules/global-require.js' {
   declare module.exports: $Exports<'eslint/lib/rules/global-require'>;
@@ -2257,6 +2286,9 @@ declare module 'eslint/lib/testers/event-generator-tester.js' {
 declare module 'eslint/lib/testers/rule-tester.js' {
   declare module.exports: $Exports<'eslint/lib/testers/rule-tester'>;
 }
+declare module 'eslint/lib/testers/test-parser.js' {
+  declare module.exports: $Exports<'eslint/lib/testers/test-parser'>;
+}
 declare module 'eslint/lib/timing.js' {
   declare module.exports: $Exports<'eslint/lib/timing'>;
 }
@@ -2298,6 +2330,9 @@ declare module 'eslint/lib/token-store/skip-cursor.js' {
 }
 declare module 'eslint/lib/token-store/utils.js' {
   declare module.exports: $Exports<'eslint/lib/token-store/utils'>;
+}
+declare module 'eslint/lib/util/ajv.js' {
+  declare module.exports: $Exports<'eslint/lib/util/ajv'>;
 }
 declare module 'eslint/lib/util/fix-tracker.js' {
   declare module.exports: $Exports<'eslint/lib/util/fix-tracker'>;
