@@ -1,5 +1,5 @@
-// flow-typed signature: 49d74be53c39f3bab40c4c02ddb7cc18
-// flow-typed version: <<STUB>>/flow-typed_v^2.1.2/flow_v0.48.0
+// flow-typed signature: d9b2436b1c1ea4aa8ccdbe43fa9c95af
+// flow-typed version: <<STUB>>/flow-typed_v^2.1.2/flow_v0.53.1
 
 /**
  * This is an autogenerated libdef stub for:
