@@ -1,5 +1,5 @@
-// flow-typed signature: 485ceb406a3c65551d82effe4b9a9392
-// flow-typed version: <<STUB>>/glob_v^7.1.2/flow_v0.48.0
+// flow-typed signature: 8a78b2258745a8845d2c07f9ed3ffa5c
+// flow-typed version: <<STUB>>/glob_v^7.1.2/flow_v0.53.1
 
 /**
  * This is an autogenerated libdef stub for:
