@@ -1,5 +1,5 @@
-// flow-typed signature: 7fe75a496d83d80512fdd722c308f7ec
-// flow-typed version: <<STUB>>/jasmine-reporters_v^2.2.1/flow_v0.48.0
+// flow-typed signature: cdfdb77833b7a63c814738410394baa2
+// flow-typed version: <<STUB>>/jasmine-reporters_v^2.2.1/flow_v0.53.1
 
 /**
  * This is an autogenerated libdef stub for:
