@@ -1,5 +1,5 @@
-// flow-typed signature: 77db0a7f13a9ee25d460ed6216518e70
-// flow-typed version: <<STUB>>/lerna_v^2.0.0-rc.5/flow_v0.48.0
+// flow-typed signature: e49a5e9920aeb82daeaf4b6c0b2ed973
+// flow-typed version: <<STUB>>/lerna_v^2.0.0/flow_v0.53.1
 
 /**
  * This is an autogenerated libdef stub for:
@@ -82,10 +82,6 @@ declare module 'lerna/lib/ConventionalCommitUtilities' {
   declare module.exports: any;
 }
 
-declare module 'lerna/lib/ExitHandler' {
-  declare module.exports: any;
-}
-
 declare module 'lerna/lib/FileSystemUtilities' {
   declare module.exports: any;
 }
@@ -95,10 +91,6 @@ declare module 'lerna/lib/GitUtilities' {
 }
 
 declare module 'lerna/lib/index' {
-  declare module.exports: any;
-}
-
-declare module 'lerna/lib/logger' {
   declare module.exports: any;
 }
 
@@ -115,10 +107,6 @@ declare module 'lerna/lib/PackageGraph' {
 }
 
 declare module 'lerna/lib/PackageUtilities' {
-  declare module.exports: any;
-}
-
-declare module 'lerna/lib/progressBar' {
   declare module.exports: any;
 }
 
@@ -196,9 +184,6 @@ declare module 'lerna/lib/commands/UpdatedCommand.js' {
 declare module 'lerna/lib/ConventionalCommitUtilities.js' {
   declare module.exports: $Exports<'lerna/lib/ConventionalCommitUtilities'>;
 }
-declare module 'lerna/lib/ExitHandler.js' {
-  declare module.exports: $Exports<'lerna/lib/ExitHandler'>;
-}
 declare module 'lerna/lib/FileSystemUtilities.js' {
   declare module.exports: $Exports<'lerna/lib/FileSystemUtilities'>;
 }
@@ -207,9 +192,6 @@ declare module 'lerna/lib/GitUtilities.js' {
 }
 declare module 'lerna/lib/index.js' {
   declare module.exports: $Exports<'lerna/lib/index'>;
-}
-declare module 'lerna/lib/logger.js' {
-  declare module.exports: $Exports<'lerna/lib/logger'>;
 }
 declare module 'lerna/lib/NpmUtilities.js' {
   declare module.exports: $Exports<'lerna/lib/NpmUtilities'>;
@@ -222,9 +204,6 @@ declare module 'lerna/lib/PackageGraph.js' {
 }
 declare module 'lerna/lib/PackageUtilities.js' {
   declare module.exports: $Exports<'lerna/lib/PackageUtilities'>;
-}
-declare module 'lerna/lib/progressBar.js' {
-  declare module.exports: $Exports<'lerna/lib/progressBar'>;
 }
 declare module 'lerna/lib/PromptUtilities.js' {
   declare module.exports: $Exports<'lerna/lib/PromptUtilities'>;
