@@ -1,5 +1,5 @@
-// flow-typed signature: 21810ae4a579e16d9e695bd1ad0a5679
-// flow-typed version: <<STUB>>/node-emoji_v^1.5.1/flow_v0.48.0
+// flow-typed signature: d0aac4f08e2fb24853db9e56e97186cd
+// flow-typed version: <<STUB>>/node-emoji_v^1.6.1/flow_v0.53.1
 
 /**
  * This is an autogenerated libdef stub for:
