@@ -1,5 +1,5 @@
-// flow-typed signature: 933a5cb9cea3011f0285f006a63d791a
-// flow-typed version: <<STUB>>/react-test-renderer_v^15.6.1/flow_v0.48.0
+// flow-typed signature: 90a58e24496065251f481f1a0179049c
+// flow-typed version: <<STUB>>/react-test-renderer_v^15.6.1/flow_v0.53.1
 
 /**
  * This is an autogenerated libdef stub for:
