@@ -1,5 +1,5 @@
-// flow-typed signature: 3ebb7dc9640f087ab46c79a1c365e10a
-// flow-typed version: <<STUB>>/rimraf_v^2.6.1/flow_v0.48.0
+// flow-typed signature: 036fbdc0146afcb08f2e0fc0a7dde33b
+// flow-typed version: <<STUB>>/rimraf_v^2.6.1/flow_v0.53.1
 
 /**
  * This is an autogenerated libdef stub for:
