@@ -1,5 +1,5 @@
-// flow-typed signature: d90f90b49c011f858b9655d9557b70b1
-// flow-typed version: <<STUB>>/rollup-plugin-babel_v^2.7.1/flow_v0.48.0
+// flow-typed signature: ab5e2931814a32160f6fb4e688f5e1a2
+// flow-typed version: <<STUB>>/rollup-plugin-babel_v^3.0.2/flow_v0.53.1
 
 /**
  * This is an autogenerated libdef stub for:
@@ -34,6 +34,10 @@ declare module 'rollup-plugin-babel/src/constants' {
   declare module.exports: any;
 }
 
+declare module 'rollup-plugin-babel/src/helperPlugin' {
+  declare module.exports: any;
+}
+
 declare module 'rollup-plugin-babel/src/index' {
   declare module.exports: any;
 }
@@ -55,6 +59,9 @@ declare module 'rollup-plugin-babel/dist/rollup-plugin-babel.es.js' {
 }
 declare module 'rollup-plugin-babel/src/constants.js' {
   declare module.exports: $Exports<'rollup-plugin-babel/src/constants'>;
+}
+declare module 'rollup-plugin-babel/src/helperPlugin.js' {
+  declare module.exports: $Exports<'rollup-plugin-babel/src/helperPlugin'>;
 }
 declare module 'rollup-plugin-babel/src/index.js' {
   declare module.exports: $Exports<'rollup-plugin-babel/src/index'>;
