@@ -1,5 +1,5 @@
-// flow-typed signature: 1be73e98115576db175d62db03d1601e
-// flow-typed version: <<STUB>>/rollup_v^0.43.0/flow_v0.48.0
+// flow-typed signature: 52267340eb2fb77959fc62b699ae54cf
+// flow-typed version: <<STUB>>/rollup_v^0.48.2/flow_v0.53.1
 
 /**
  * This is an autogenerated libdef stub for:
