@@ -20,7 +20,7 @@ describe('WithI18n', function () {
   // Pass all props to spy on render
   const sinkFactory = (options = {}) => {
     const spy = jest.fn()
-    const Sink = WithI18n(options)(class extends React.Component {
+    const Sink = WithI18n(options)(class extends React.Component<*> {
       customMethod = () => 42
 
       render () {
@@ -124,17 +124,13 @@ describe('WithI18n', function () {
   it('should hold ref to wrapped instance when withRef is enabled', function () {
     const { node } = mountHoc({}, { withRef: true })
     const hoc = node.getNode()
-    // $FlowIgnore: getWrappedInstance() is custom method of WithI18n
     expect(hoc.getWrappedInstance()).not.toBeNull()
-    // $FlowIgnore: getWrappedInstance() is custom method of WithI18n
     expect(hoc.getWrappedInstance().customMethod).not.toBeNull()
-    // $FlowIgnore: getWrappedInstance() is custom method of WithI18n
     expect(hoc.getWrappedInstance().customMethod()).toEqual(42)
   })
 
   it('should not hold ref to wrapped instance when withRef is disabled', function () {
     const { node } = mountHoc()
-    // $FlowIgnore: getWrappedInstance() is custom method of WithI18n
     expect(() => node.getNode().getWrappedInstance()).toThrow(
       'To access the wrapped instance, you need to specify { withRef: true } in the options argument of the withI18n() call.'
     )
