@@ -12,10 +12,14 @@ type FormatProps<V, FormatOptions> = {
   }
 } & RenderProps
 
-function createFormat<V, FormatOptions, P: FormatProps<V, FormatOptions>> (formatFunction: (language: string, format?: FormatOptions) => (value: V) => string): ReactClass<P> {
+function createFormat<V, FormatOptions> (formatFunction: (language: string, format?: FormatOptions) => (value: V) => string) {
   return function ({ value, format, i18n, className, render }: FormatProps<V, FormatOptions>) {
     const formatter = formatFunction(i18n.language, format)
-    return <Render className={className} render={render}>{formatter(value)}</Render>
+    return <Render
+      className={className}
+      render={render}
+      value={formatter(value)}
+    />
   }
 }
 
